requirements.txt
pya2ldb
chardet